added a readme
