+++ TeamCode/src/main/java/pedroPathing/readme.md
@@ -1,0 +1,22 @@
+# App structure cheatsheet
+
+## actuators
+- High-Level control on robot hardware
+
+## examples
+- PedroPathing code examples
+
+## gamepad 
+- High-Level gamepad control
+
+## PPConstants
+- PedroPathing follower constants
+
+## programs
+- TeleOp and Autonomous classes for competition
+
+## simulation
+- Not yet ready
+
+## testers
+- Debugging TeleOps: full robot control
